fix: (2) rename with deleting case
